Controller CRUD 로직 구현 & 업테이트 및 GET 서비스 추가 구현

+++ src/main/java/com/fastcampus/board/dto/PostDto.java
@@ -10,7 +10,7 @@ import lombok.*;
 public class PostDto {
     // PostDto는 JSP 화면과 Spring 서버가 주고받는 객체입니다.
     // PostDto를 수정하시기보다는 Post 클래스를 만들어 사용하기를 권장합니다.
-    private long postId;
+    private int postId;
     private String nickName;
     private String title;
     private String content;
